Increased version number of NativeInterpreter and dependent plugins.
