+++ HeuristicLab.Clients.Hive.Slave/3.3/Plugin.cs.frame
@@ -28,8 +28,7 @@ using HeuristicLab.PluginInfrastructure;
 
 namespace HeuristicLab.Clients.Hive.SlaveCore {
   [Plugin("HeuristicLab.Clients.Hive.SlaveCore", "3.3.16.$WCREV$")]
-  [PluginFile("HeuristicLab.Clients.Hive.SlaveCore-3.3.dll", PluginFileType.Assembly)]
-  
+  [PluginFile("HeuristicLab.Clients.Hive.SlaveCore-3.3.dll", PluginFileType.Assembly)]  
   [PluginDependency("HeuristicLab.Clients.Common", "3.3")]
   [PluginDependency("HeuristicLab.Clients.Hive", "3.3")]
   [PluginDependency("HeuristicLab.Common", "3.3")]
